requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
MarkupSafe==2.1.5
openai==1.50.2
outcome==1.3.0.post0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
